fix(helpers): refactor blocksToMarkdown into modular renderer
Extract Blocks-to-Markdown rendering logic from src/tools/helpers/markdown.ts into a new dedicated file src/tools/helpers/markdown-renderer.ts. This addresses the "Long Function: blocksToMarkdown" issue by separating parsing and rendering concerns and significantly reducing the size of the primary markdown helper file.

Maintains backward compatibility by re-exporting rendering functions from the original module.

Testing:
- Ran bun run test src/tools/helpers/markdown.test.ts (182 tests passed)
- Verified coverage (91%+ on lines)
- Ran bun run check:fix for linting and type checking.
